feat(groups): moved groups index to redis set
The in-memory implementation of groups would not work with multiple
processes. If you are using this module it's for scaling and should
be running multiple node processes anyways.

These changes won't be compatible with the memory cache, but as
mentioned above the memory cache won't support cache invalidation
with multiple processes anyways.

* fixed some logging typos
* moved clearing of a group to it's own function for simplicity